requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.3
dnspython==2.8.0
email-validator==2.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pydantic==2.12.0
pydantic_core==2.41.1
python-dotenv==1.0.1
python-telegram-bot==21.3
requests==2.31.0
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
